Make some preparations for moving to syntax-color-lib
Make hash-lang.rkt itself less idiosyncratic to how we want to use it
in Emacs.

Instead of supplying the class an async-channel, now it takes an
optional on-notify procedure. Something using the class only for
indent might not need changed-token notifications at all. And even
something that needs them can choose how to handle them (although some
non-blocking technique like an async-channel is important).

Similarly, move the code that massaged the notifications into the
desired format for Emacs, from hash-lang.rkt to hash-lang-bridge.rkt.
By supplying paren-matches as the first argument to on-notify, we
give it the ability to massage parenthesis tokens. Also, since
on-notify gets the full token struct, it can handle the case where
token-type is a hash-table instead of a symbol, as with module-lexer*.

Although I'm not sure this is everything that needs to be done, or
that all of these details are just right, it's at least a first step.