Various changes
Back end:

Handle case of inserting in middle of existing token; split it before
expanding the interval-map to make the old/new compare work correctly
and emit minimal changed-token notifications.

Add a message when racket-hash-lang-mode is used with a #lang that
supplies nothing special beyond a color-lexer -- no grouping-position
or indent functions. In that case plain old racket-mode would suffice,
unless a user prefers the "more boring" (or "less garish", depending
on your opinion) coloring. Similarly append symbols to the mode-line
lighter when each of these is supplied by a lang. (Although I'm not
sure this is really the right UX, it's information that at least is
helpful to me now when dogfooding and trying to understand what is
supposed to work and why.)

Front end:

Stop setting forward-exp-function. Instead supply navigation commands.
When the lang supplies no grouping-position at all, or when it does
but a specific call returns "use s-expression", we use the default
Emacs commands.

Switch from cl-case to pcase.