Make textoid methods as fast as color-text%
This results from both:

1. Using paren-tree% for things like backward/forward matching. As a
result, textoid methods now look much more like those from color:text%
as opposed to the "expeditor" implementations.

2. Not taking semaphores unnecessarily.

The textoid% methods used to call token-ref, which took a semaphore
guarding access to the token-tree. Also, token-ref always retrieved
both the bounds and the data (which was "free" with an interval-map,
but requires 2 or 3 method calls with token-tree%).

We need semaphores because splay trees are blatantly not thread safe:
You do a search! to move a node to the root, then use get-root-xyz
methods. Obviously these need to be grouped and guarded for
concurrency.

What we can do is take the tokens and parens semaphores more
selectively -- only in the "outer" navigation/indent methods, where we
give ourselves as a textoid argument to a drracket:xyz function so it
may use our textoid methods. In other words, we take the semaphores
once, for the dynamic extent of those drracket functions, which may
call our textoid% methods many times; those individual methods need
not be guarded.

As before, we proceed only after block-until-updated-thru returns,
i.e. the updater thread has progressed far enough.

We do still take the tokens and parens semaphores on our updater
thread, before modifying the tokens or parens trees.