Change approach to hash-lang for REPL
Stop doing something with an offset to the last input in the buffer.

Instead, have the hash-lang handle the entire buffer. To avoid
problems lexing hopelessly "random" output, give the hash-lang a view
of the source where non-input is changed to whitespace --- everything
but newlines changed to spaces (newlines preserved so that indent
works well for input that is offset due to REPL prompts.)

So a REPL buffer with this mix of output and input:

 ";Welcome to Racket\nprompt> 42\n42\nprompt> "

will look to the hash-lang like

 "                  \n        42\n  \n        "

As a result, the hash-lang can lex the user input for purposes of
font-lock and indent, without possibly barfing on segments of user
program output and REPL prompts.

Although this is a little weird, it lets us font-lock all input in the
buffer (not just the last/tail input), and avoids littering so much
code with offset math.