Change open-input-text
Don't set! the `start` argument in the `read-in` proc because it is
used to set the initial port location.

Although I think this is N/A from a timing point of view (and indeed I
looked at this while tracking down a bug, which turned out /not/ to be
caused by this), it seems "more correct".

Also I rewrote multiple let/if stanzas using match, and added a simple
smoke test.